requirements.txt
streamlit
pandas
matplotlib
seaborn
python-pptx
openpyxl
requests
